Refactor internal class exceptions extending ParserException into new package Let exception thrown from `MTScript.raiseException` be caught by macros
