+++ src/main/java/net/rptools/maptool/client/MapToolLineParser.java
@@ -22,9 +22,7 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import net.rptools.common.expression.Result;
 import net.rptools.maptool.client.functions.*;
-import net.rptools.maptool.client.functions.AbortFunction.AbortFunctionException;
-import net.rptools.maptool.client.functions.AssertFunction.AssertFunctionException;
-import net.rptools.maptool.client.functions.ReturnFunction.ReturnFunctionException;
+import net.rptools.maptool.client.functions.exceptions.*;
 import net.rptools.maptool.client.functions.json.JSONMacroFunctions;
 import net.rptools.maptool.client.ui.htmlframe.HTMLFrameFactory;
 import net.rptools.maptool.client.ui.htmlframe.HTMLFrameFactory.FrameType;
