+++ src/main/java/net/rptools/maptool/client/functions/MacroJavaScriptBridge.java
@@ -30,7 +30,7 @@ import net.rptools.parser.VariableResolver;
 import net.rptools.parser.function.AbstractFunction;
 import org.graalvm.polyglot.*;
 
-public class MacroJavaScriptBridge extends AbstractFunction {
+public class MacroJavaScriptBridge extends AbstractFunction implements DefinesSpecialVariables {
   private static final Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
   private static final MacroJavaScriptBridge instance = new MacroJavaScriptBridge();
@@ -221,5 +221,10 @@ public class MacroJavaScriptBridge extends AbstractFunction {
     } else {
       return callingArgsStack.peek();
     }
+  }
+
+  @Override
+  public String[] getSpecialVariables() {
+    return new String[] {"macro.catchAssert"};
   }
 }
