+++ src/main/java/net/rptools/maptool/client/functions/ReturnFunction.java
@@ -16,6 +16,7 @@ package net.rptools.maptool.client.functions;
 
 import java.math.BigDecimal;
 import java.util.List;
+import net.rptools.maptool.client.functions.exceptions.*;
 import net.rptools.parser.Parser;
 import net.rptools.parser.ParserException;
 import net.rptools.parser.VariableResolver;
@@ -71,28 +72,6 @@ public class ReturnFunction extends AbstractFunction implements DefinesSpecialVa
       arrobject[1] = BigDecimal.class.getName();
       throw new ParameterException(
           String.format("Illegal argument type %s, expecting %s", arrobject));
-    }
-  }
-
-  /**
-   * Exception type thrown by return() function. Semantics are to silently halt the current
-   * execution.
-   */
-  public static class ReturnFunctionException extends ParserException {
-
-    private Object result;
-
-    public Object getResult() {
-      return result;
-    }
-
-    public void setResult(Object result) {
-      this.result = result;
-    }
-
-    public ReturnFunctionException(Object result) {
-      super("");
-      this.result = result;
     }
   }
