+++ src/main/java/net/rptools/maptool/client/macro/MacroManager.java
@@ -19,10 +19,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.MapToolMacroContext;
-import net.rptools.maptool.client.functions.AbortFunction;
 import net.rptools.maptool.client.functions.AboutMacro;
-import net.rptools.maptool.client.functions.AssertFunction;
-import net.rptools.maptool.client.functions.ReturnFunction;
+import net.rptools.maptool.client.functions.exceptions.*;
 import net.rptools.maptool.client.macro.impl.*;
 import net.rptools.maptool.client.ui.MapToolFrame;
 import net.rptools.maptool.client.ui.commandpanel.CommandPanel;
@@ -198,10 +196,10 @@ public class MacroManager {
         context.addTransform(command);
         continue;
       }
-    } catch (AbortFunction.AbortFunctionException | ReturnFunction.ReturnFunctionException fe) {
+    } catch (AbortFunctionException | ReturnFunctionException fe) {
       // Do nothing, just silently exit
       return;
-    } catch (AssertFunction.AssertFunctionException afe) {
+    } catch (JavascriptFunctionException | AssertFunctionException afe) {
       MapTool.addLocalMessage(afe.getMessage());
       return;
     } catch (ParserException e) {
